[MTS_MDOT_F405RG] reorg hal folder
- reorganisation of this target needed some extension of serial_api.c
used by all F4xx tagets.
- add arch_max to travis_build
